% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serializeDS.R
\name{semiOPBART_ecdfFromSerialize}
\alias{semiOPBART_ecdfFromSerialize}
\title{Reconstruct a normalization closure from semiOPBART_ecdfToSerialize()'s output.}
\usage{
semiOPBART_ecdfFromSerialize(Serialize_str)
}
\description{
Reconstruct a normalization closure from semiOPBART_ecdfToSerialize()'s output.
}
