% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serializeDS.R
\name{semiOPBART_treesFromSerialize}
\alias{semiOPBART_treesFromSerialize}
\title{Decode a Serialize tree bundle and load it into `forest` at `tree_idx`
(overwriting whatever was previously at those indices).}
\usage{
semiOPBART_treesFromSerialize(forest, Serialize_str, tree_idx)
}
\description{
Decode a Serialize tree bundle and load it into `forest` at `tree_idx`
(overwriting whatever was previously at those indices).
}
